+++ test/test_endpoints.py
@@ -14,6 +14,7 @@ from twisted.internet.endpoints import TCP4ServerEndpoint
 from twisted.internet.endpoints import serverFromString
 from twisted.internet.endpoints import clientFromString
 from twisted.python.failure import Failure
+from twisted.internet.error import ConnectionRefusedError
 from twisted.internet.interfaces import IReactorCore
 from twisted.internet.interfaces import IProtocolFactory
 from twisted.internet.interfaces import IProtocol
@@ -26,14 +27,19 @@ from txtorcon import ITorControlProtocol
 from txtorcon import TorConfig
 from txtorcon import launch_tor
 from txtorcon import TCPHiddenServiceEndpoint
+from txtorcon import TorClientEndpoint
 from txtorcon import TorNotFound
 from txtorcon import TCPHiddenServiceEndpointParser
 from txtorcon import IProgressProvider
 from txtorcon import TorOnionAddress
 from txtorcon.util import NoOpProtocolFactory
 from txtorcon.endpoints import get_global_tor                       # FIXME
+from txtorcon.endpoints import default_tcp4_endpoint_generator
 
 import util
+
+
+connectionRefusedFailure = Failure(ConnectionRefusedError())
 
 
 class EndpointTests(unittest.TestCase):
@@ -522,3 +528,157 @@ class FakeReactorTcp(FakeReactor):
             print "BLAAAAAM", args
         r.connect = blam
         return r
+
+
+class FakeTorSocksEndpoint(object):
+    def __init__(self, *args, **kw):
+        self.host     = args[1]
+        self.port     = args[2]
+        self.transport = None
+
+        if kw.has_key('failure'):
+            self.failure = kw['failure']
+        else:
+            self.failure = None
+        if kw.has_key('acceptPort'):
+            self.acceptPort = kw['acceptPort']
+        else:
+            self.acceptPort = None
+
+    def connect(self, fac):
+        self.factory = fac
+        if self.acceptPort:
+            if self.port != self.acceptPort:
+                return defer.fail(self.failure)
+        else:
+            if self.failure:
+                return defer.fail(self.failure)
+        self.proto = fac.buildProtocol(None)
+        transport = proto_helpers.StringTransport()
+        self.proto.makeConnection(transport)
+        self.transport = transport
+        return defer.succeed(self.proto)
+
+
+class TestTorClientEndpoint(unittest.TestCase):
+
+    def test_client_connection_failed(self):
+        """
+        This test is equivalent to txsocksx's
+        TestSOCKS4ClientEndpoint.test_clientConnectionFailed
+        """
+        def FailTorSocksEndpointGenerator(*args, **kw):
+            kw['failure'] = connectionRefusedFailure
+            return FakeTorSocksEndpoint(*args, **kw)
+        endpoint = TorClientEndpoint('', 0, proxyEndpointGenerator=FailTorSocksEndpointGenerator)
+        d = endpoint.connect(None)
+        return self.assertFailure(d, ConnectionRefusedError)
+
+    def test_client_connection_failed_user_password(self):
+        """
+        Same as above, but with a username/password.
+        """
+        def FailTorSocksEndpointGenerator(*args, **kw):
+            kw['failure'] = connectionRefusedFailure
+            return FakeTorSocksEndpoint(*args, **kw)
+        endpoint = TorClientEndpoint(
+            'invalid host', 0,
+            socks_username='billy', socks_password='s333cure',
+            proxyEndpointGenerator=FailTorSocksEndpointGenerator)
+        d = endpoint.connect(None)
+        return self.assertFailure(d, ConnectionRefusedError)
+
+    def test_default_generator(self):
+        # just ensuring the default generator doesn't blow updoesn't blow up
+        default_tcp4_endpoint_generator(None, 'foo.bar', 1234)
+
+    def test_no_host(self):
+        self.assertRaises(
+            ValueError,
+            TorClientEndpoint, None, None
+        )
+
+    def test_parser_basic(self):
+        ep = clientFromString(None, 'tor:host=timaq4ygg2iegci7.onion:port=80:socksPort=9050')
+
+        self.assertEqual(ep.host, 'timaq4ygg2iegci7.onion')
+        self.assertEqual(ep.port, 80)
+        self.assertEqual(ep.socks_port, 9050)
+
+    def test_parser_user_password(self):
+        epstring = 'tor:host=torproject.org:port=443' + \
+                   ':socksUsername=foo:socksPassword=bar'
+        ep = clientFromString(None, epstring)
+
+        self.assertEqual(ep.host, 'torproject.org')
+        self.assertEqual(ep.port, 443)
+        self.assertEqual(ep.socks_username, 'foo')
+        self.assertEqual(ep.socks_password, 'bar')
+
+    def test_default_factory(self):
+        """
+        This test is equivalent to txsocksx's TestSOCKS5ClientEndpoint.test_defaultFactory
+        """
+        def TorSocksEndpointGenerator(*args, **kw):
+            return FakeTorSocksEndpoint(*args, **kw)
+        endpoint = TorClientEndpoint('', 0, proxyEndpointGenerator=TorSocksEndpointGenerator)
+        endpoint.connect(None)
+        self.assertEqual(endpoint.torSocksEndpoint.transport.value(), '\x05\x01\x00')
+
+    def test_good_port_retry(self):
+        """
+        This tests that our Tor client endpoint retry logic works correctly.
+        We create a proxy endpoint that fires a connectionRefusedFailure
+        unless the connecting port matches. We attempt to connect with the
+        proxy endpoint for each port that the Tor client endpoint will try.
+        """
+        success_ports = TorClientEndpoint.socks_ports_to_try
+        for port in success_ports:
+            def TorSocksEndpointGenerator(*args, **kw):
+                kw['acceptPort'] = port
+                kw['failure']    = connectionRefusedFailure
+                return FakeTorSocksEndpoint(*args, **kw)
+            endpoint = TorClientEndpoint('', 0, proxyEndpointGenerator=TorSocksEndpointGenerator)
+            endpoint.connect(None)
+            self.assertEqual(endpoint.torSocksEndpoint.transport.value(), '\x05\x01\x00')
+
+    def test_bad_port_retry(self):
+        """
+        This tests failure to connect to the ports on the "try" list.
+        """
+        fail_ports = [1984, 666]
+        for port in fail_ports:
+            def TorSocksEndpointGenerator(*args, **kw):
+                kw['acceptPort'] = port
+                kw['failure']    = connectionRefusedFailure
+                return FakeTorSocksEndpoint(*args, **kw)
+            endpoint = TorClientEndpoint('', 0, proxyEndpointGenerator=TorSocksEndpointGenerator)
+            d = endpoint.connect(None)
+            return self.assertFailure(d, ConnectionRefusedError)
+
+    def test_good_no_guess_socks_port(self):
+        """
+        This tests that if a SOCKS port is specified, we *only* attempt to
+        connect to that SOCKS port.
+        """
+        def TorSocksEndpointGenerator(*args, **kw):
+            kw['acceptPort'] = 6669
+            kw['failure']    = connectionRefusedFailure
+            return FakeTorSocksEndpoint(*args, **kw)
+        endpoint = TorClientEndpoint('', 0, proxyEndpointGenerator=TorSocksEndpointGenerator, socks_port=6669)
+        endpoint.connect(None)
+        self.assertEqual(endpoint.torSocksEndpoint.transport.value(), '\x05\x01\x00')
+
+    def test_bad_no_guess_socks_port(self):
+        """
+        This tests that are connection fails if we try to connect to an unavailable
+        specified SOCKS port... even if there is a valid SOCKS port listening on
+        the socks_ports_to_try list.
+        """
+        def TorSocksEndpointGenerator(*args, **kw):
+            kw['acceptPort'] = 9050
+            kw['failure']    = connectionRefusedFailure
+            return FakeTorSocksEndpoint(*args, **kw)
+        endpoint = TorClientEndpoint('', 0, proxyEndpointGenerator=TorSocksEndpointGenerator, socks_port=6669)
+        d = endpoint.connect(None)
+        self.assertFailure(d, ConnectionRefusedError)